сделал публикацию по нажатию на кнопку на посте

+++ Bot/src/routerBot/instantPublicationPosts.ts
@@ -1,42 +1,44 @@
 import { bot } from './index.js';
 import fs from 'fs';
+import { waterMark } from '../const/const.js';
 
-export async function instantPublicationPosts(files: any) {
-  const media: any = [];
-  for (const file of files) {
-    const mediaBlock = {
-      type: 'photo',
-      media: fs.readFileSync(file.path),
-      caption: '#QugorArts'
-    };
-    media.push(mediaBlock);
+export async function instantPublicationPosts(files: any, stringArray?: boolean) {
+  return new Promise((resolve, reject) => {
+    const media: any = [];
+    for (const file of files) {
+      const mediaBlock = {
+        type: 'photo',
+        media: stringArray ? file : fs.readFileSync(file.path),
+        caption: waterMark
+      };
+      media.push(mediaBlock);
 
-    fs.unlink(`${file.destination}${file.filename}`, (err) => {
-      if (err) {
-        console.error('Ошибка при удалении файла:', err);
-      } else {
-        console.log('Файл успешно удален:', file.path);
-      }
-    });
-  }
-
-  if (media.length > 10) {
-    console.error('Слишком много медиафайлов');
-    return;
-  }
-
-  if (process.env.CHAT_ID_DEV) {
-    try {
-      bot.sendMediaGroup(process.env.CHAT_ID_DEV, media)
-        .then(async () => {
-          // console.log()
-        })
-        .catch((error: Error) => {
-          console.error('Ошибка при отправке сообщения:', error);
+      if (!stringArray) {
+        fs.unlink(`${file.destination}${file.filename}`, (err) => {
+          if (err) {
+            console.error('Ошибка при удалении файла:', err);
+            reject(err);
+          } else {
+            console.log('Файл успешно удален:', file.path);
+          }
         });
-    } catch (e) {
-      console.log(e);
+      }
     }
-  }
+    console.log(media);
+    if (media.length > 10) {
+      console.error('Слишком много медиафайлов');
+      reject(new Error('Слишком много медиафайлов'));
+      return;
+    }
 
+    bot.sendMediaGroup(process.env.CHAT_ID_DEV, media)
+      .then(() => {
+        resolve(true);
+      })
+      .catch((error: Error) => {
+        console.error('Ошибка при отправке сообщения:', error);
+        reject(error);
+      });
+  });
 }
+
